+++ arend-lib/meta/src/main/java/org/arend/lib/meta/ExtMeta.java
@@ -705,7 +705,7 @@ public class ExtMeta extends BaseMetaDefinition {
   private static boolean sortAtLeast(CoreSortExpression sort, BigInteger level) {
     if (sort instanceof ConstSortExpression constSort) {
       ConstLevel hLevel = constSort.getSort().getHLevel();
-      return hLevel.isInfinity() || BigInteger.valueOf(hLevel.value()).compareTo(level) >= 0;
+      return hLevel.isInfinity() || hLevel.value().compareTo(level) >= 0;
     } else if (sort instanceof PreviousSortExpression equalitySort) {
       return sortAtLeast(equalitySort.getSort(), level.add(BigInteger.ONE));
     } else if (sort instanceof MaxSortExpression maxSort) {
