+++ base/src/main/java/org/arend/term/concrete/ConcreteExpressionFactory.java
@@ -217,7 +217,7 @@ public class ConcreteExpressionFactory {
     return new Concrete.UniverseExpression(null, new Concrete.NumberLevelExpression(null, BigInteger.valueOf(level)), null, ConcreteUniverseExpression.Kind.TYPE);
   }
 
-  public static Concrete.UniverseExpression cUniverse(Concrete.LevelExpression pLevel, Integer hLevel, Concrete.UniverseExpression.Kind kind) {
+  public static Concrete.UniverseExpression cUniverse(Concrete.LevelExpression pLevel, BigInteger hLevel, Concrete.UniverseExpression.Kind kind) {
     return new Concrete.UniverseExpression(null, pLevel, hLevel, kind);
   }
