+++ intellij/src/main/kotlin/org/arend/refactoring/ArendRefactoringUtils.kt
@@ -837,7 +837,7 @@ private object PrecVisitor : AbstractExpressionVisitor<Void?, Int> {
     override fun visitReference(data: Any?, referent: Referable, fixity: Fixity?, pLevels: Collection<Abstract.LevelExpression>?, params: Void?) =
         if (pLevels != null) APP_PREC else MAX_PREC
 
-    override fun visitUniverse(data: Any?, pLevelNum: BigInteger?, hLevelNum: Int?, pLevel: Abstract.LevelExpression?, params: Void?) =
+    override fun visitUniverse(data: Any?, pLevelNum: BigInteger?, hLevelNum: BigInteger?, pLevel: Abstract.LevelExpression?, params: Void?) =
         if (pLevel != null) APP_PREC else MAX_PREC
 
     override fun visitCatUniverse(data: Any?, pLevelNum: BigInteger?, pLevel: Abstract.LevelExpression?, params: Void?): Int? =
